backend/pyproject.toml
[project]
name = "declaro-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "simpleeval>=0.9.13",
  "openpyxl>=3.1.0",
  "python-multipart>=0.0.9",
  "psycopg2-binary>=2.9.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "slowapi>=0.1.9",
  "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "pytest-mock>=3.14.0",
]
